Correct video timing and sprite collection
The per-cycle update was being called once per m-cycle, but the video
emulation needs to be updated once per clock (4x faster)! Renamed the
"per clock cycle updates" chunk to "per machine cycle updates" to
clarify when it is called.

The sprite collection loop could write out of bounds when shifting the
last sprite the array out. Corrected the logic here to make it clearer
and hopdfully correct.